requirements.txt
Flask==2.0.1
jsonpickle==2.0.0
opencv-python==4.5.2.52
tensorflow==2.5.0
tensorflow-serving-api==2.5.1